qr-code-generator/requirements.txt
asgiref==3.8.1
Django==5.2.3
numpy==2.3.1
opencv-python==4.11.0.86
pillow==11.2.1
pyzbar==0.1.9
qrcode==8.2
sqlparse==0.5.3
